requirements/local.txt
-r base.txt
django-braces==1.9.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.5
django-material==0.8.0
django-model-utils==2.5.2
django-star-ratings==0.5.3
django-stronghold==0.2.8
django-widget-tweaks==1.4.1
isort==4.2.5
Pillow==3.3.1
sqlparse==0.2.1
